Develop

+++ Controller/users.php
@@ -10,7 +10,15 @@ if(!empty($_SERVER['HTTP_X_REQUESTED_WITH']) && $_SERVER['HTTP_X_REQUESTED_WITH'
     
     if (isset($_GET['action']) && $_GET['action'] === 'toggle' && isset($_GET['id']) && is_numeric($_GET['id'])) {
         $id = cleanString($_GET['id']);
-        $result = toggle_enabled($pdo, $id);
+        if ($id == $_SESSION['user_id']) {
+            logAction($pdo, 'toggle_user_status', "Failed to deActivate/activate user ID: $id by admin: " . $_SESSION['user_id']);
+            header('Content-Type: application/json', true, 403);
+            echo json_encode(['status' => 'error', 'message' => 'You cannot change your own account status.']);
+            exit();
+        } else {
+            $result = toggle_enabled($pdo, $id);
+        }
+        
         if ($result) {
             logAction($pdo, 'activate/deActivate', "user $id status changed by admin : " . $_SESSION['user_id']);
             header('Content-Type: application/json');
@@ -23,15 +31,24 @@ if(!empty($_SERVER['HTTP_X_REQUESTED_WITH']) && $_SERVER['HTTP_X_REQUESTED_WITH'
         exit();
     } elseif (isset($_GET['action']) && $_GET['action'] === 'delete' && isset($_GET['id']) && is_numeric($_GET['id'])) {
         $id = cleanString($_GET['id']);
-        $result = deleteUser($pdo, $id);
+        if ($id == $_SESSION['user_id']) {
+            logAction($pdo, 'delete_user', "Failed to delete user ID: $id by admin: " . $_SESSION['user_id']);
+            header('Content-Type: application/json', true, 403);
+            echo json_encode(['status' => 'error', 'message' => 'You cannot delete your own account.']);
+            exit();
+        } else {
+            $result = deleteUser($pdo, $id);
+        }
+        
         if ($result > 0) {
             logAction($pdo, 'delete_user', "user $id deleted by admin: " . $_SESSION['user_id']);
             header('Content-Type: application/json');
             echo json_encode(['status' => 'success', 'message' => 'User deleted successfully.']);
         } else {
+
             logAction($pdo, 'delete_user', "Failed to delete user ID: $id by admin: " . $_SESSION['user_id']);
             header('Content-Type: application/json', true, 500);
-            echo json_encode(['status' => 'error', 'message' => 'Failed to delete user.']);
+            echo json_encode(['status' => 'error', 'message' => 'Failed to delete user.'. $result]);
         }
         exit();
     }
